debian: ensure prometheus_client is available during package build
pybuild wants to run tests during build, and even though we don't have
any, the unittest module's test scanner wants the idg_exporter package
to be importable.

This wasn't a problem when igd_exporter.__init__ was empty, but now that
it imports other modules, their dependencies must also be available.